# packages/contracts/openapi/openapi.json
{
  "components": {
    "parameters": {
...
      "Cursor": {
        "in": "query",
        "name": "cursor",
        "required": false,
        "schema": {
          "type": [
            "string",
            "null"
          ]
        }
      },
      "IdempotencyKey": {
        "in": "header",
        "name": "Idempotency-Key",
        "required": true,
        "schema": {
          "maxLength": 128,
          "minLength": 8,
          "type": "string"
        }
      },
...
    },
    "responses": {
      "Sprint2Error401": {
        "content": {
          "application/json": {
            "schema": {
              "$ref": "#/components/schemas/Sprint2ErrorResponse"
            }
          }
        },
        "description": "Sprint 2 error (401)",
        "x-error-codes": [
          "AUTH_REQUIRED"
        ],
        "x-http-status": 401
      },
      "Sprint2Error403": {
        "content": {
          "application/json": {
            "schema": {
              "$ref": "#/components/schemas/Sprint2ErrorResponse"
            }
          }
        },
        "description": "Sprint 2 error (403)",
        "x-error-codes": [
          "FORBIDDEN",
          "PERMISSION_CHANGED"
        ],
        "x-http-status": 403
      },
      "Sprint2Error404": {
        "content": {
          "application/json": {
            "schema": {
              "$ref": "#/components/schemas/Sprint2ErrorResponse"
            }
          }
        },
        "description": "Sprint 2 error (404)",
        "x-error-codes": [
          "RESOURCE_NOT_FOUND"
        ],
        "x-http-status": 404
      },
      "Sprint2Error409": {
        "content": {
          "application/json": {
            "schema": {
              "$ref": "#/components/schemas/Sprint2ErrorResponse"
            }
          }
        },
        "description": "Sprint 2 error (409)",
        "x-error-codes": [
          "VERSION_CONFLICT",
          "IDEMPOTENCY_CONFLICT",
          "CLARIFICATION_ROUND_LIMIT_REACHED",
          "DEEP_CONFIRMATION_REQUIRED",
          "AI_TASK_NOT_CANCELLABLE",
          "AI_TASK_NOT_RETRYABLE",
          "RESULT_NOT_READY",
          "RESULT_QUALITY_BLOCKED",
          "STALE_TARGET",
          "TRACEABILITY_INCOMPLETE",
          "CLARIFICATION_ROUND_INVALID"
        ],
        "x-http-status": 409
      },
      "Sprint2Error422": {
        "content": {
          "application/json": {
            "schema": {
              "$ref": "#/components/schemas/Sprint2ErrorResponse"
            }
          }
        },
        "description": "Sprint 2 error (422)",
        "x-error-codes": [
          "VALIDATION_ERROR",
          "CLARIFICATION_MODE_REASON_REQUIRED",
          "RISK_ACCEPTANCE_INVALID"
        ],
        "x-http-status": 422
      },
      "Sprint2Error429": {
        "content": {
          "application/json": {
            "schema": {
              "$ref": "#/components/schemas/Sprint2ErrorResponse"
            }
          }
        },
        "description": "Sprint 2 error (429)",
        "x-error-codes": [
          "AI_QUOTA_EXCEEDED"
        ],
        "x-http-status": 429
      },
      "Sprint2Error503": {
        "content": {
          "application/json": {
            "schema": {
              "$ref": "#/components/schemas/Sprint2ErrorResponse"
            }
          }
        },
        "description": "Sprint 2 error (503)",
        "x-error-codes": [
          "QUEUE_UNAVAILABLE",
          "DEPENDENCY_UNAVAILABLE"
        ],
        "x-http-status": 503
      },
...
    },
    "schemas": {
...
      "RequirementListResponse": {
        "additionalProperties": false,
        "properties": {
          "code": {
            "const": "OK",
            "type": "string"
          },
          "data": {
            "$ref": "#/components/schemas/RequirementListData"
          },
          "message": {
            "type": "string"
          },
          "trace_id": {
            "type": "string"
          }
        },
        "required": [
          "code",
          "message",
          "data",
          "trace_id"
        ],
        "type": "object"
      },
...
    },
    "securitySchemes": {
      "bearerAuth": {
        "bearerFormat": "JWT",
        "scheme": "bearer",
        "type": "http"
      },
      "refreshCookie": {
        "description": "Secure, HttpOnly, SameSite refresh cookie; default lifetime seven days.",
        "in": "cookie",
        "name": "refresh_token",
        "type": "apiKey"
      },
      "serviceBearerAuth": {
        "bearerFormat": "JWT",
        "description": "Internal HS256 service JWT; fixed issuer/audience/scope and short lifetime.",
        "scheme": "bearer",
        "type": "http",
        "x-ai-caller-subject": "ai-api",
        "x-required-audience": "business-api",
        "x-required-issuers": [
          "ai-api",
          "monitoring"
        ],
        "x-required-scope": "health"
      }
    }
  },
  "info": {
    "description": "Sprint 2 Requirement/Baseline R4 implementation candidate authorized by PORTFOLIO-P1-RUNTIME-01. It remains pending Review and the shared real-runtime Gate; persistence adapters and FLOW remain disabled.",
    "summary": "Sprint 2 Requirement/Baseline R4 implementation candidate",
    "title": "AI Product Design and Validation API",
    "version": "0.1.0-dev",
    "x-contract-reviewers": [
      "frontend",
      "ai-data",
      "review"
    ],
    "x-contract-status": "candidate"
  },
  "openapi": "3.1.0",
  "paths": {
...
    "/api/v1/project-versions/{version_id}/requirements": {
      "get": {
        "operationId": "listRequirements",
        "parameters": [
          {
            "in": "path",
            "name": "version_id",
            "required": true,
            "schema": {
              "type": "string"
            }
          },
          {
            "$ref": "#/components/parameters/Cursor"
          },
          {
            "in": "query",
            "name": "status",
            "required": false,
            "schema": {
              "type": "string"
            }
          }
        ],
        "responses": {
          "200": {
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/RequirementListResponse"
                }
              }
            },
            "description": "Successful response"
          },
          "401": {
            "$ref": "#/components/responses/Sprint2Error401"
          },
          "403": {
            "$ref": "#/components/responses/Sprint2Error403"
          },
          "404": {
            "$ref": "#/components/responses/Sprint2Error404"
          },
          "409": {
            "$ref": "#/components/responses/Sprint2Error409"
          },
          "422": {
            "$ref": "#/components/responses/Sprint2Error422"
          },
          "429": {
            "$ref": "#/components/responses/Sprint2Error429"
          },
          "503": {
            "$ref": "#/components/responses/Sprint2Error503"
          }
        },
        "security": [
          {
            "bearerAuth": []
          }
        ],
        "tags": [
          "requirements"
        ],
        "x-command-semantics": {
          "actor": "project_member_with_target_read_permission",
          "admin_bypass": false,
          "queue_unavailable_no_false_queued": false
        },
        "x-contract-phase": "sprint-2-p1-runtime-r4-candidate",
        "x-error-http-mapping": {
          "AI_QUOTA_EXCEEDED": 429,
          "AI_TASK_NOT_CANCELLABLE": 409,
          "AI_TASK_NOT_RETRYABLE": 409,
          "AUTH_REQUIRED": 401,
          "CLARIFICATION_MODE_REASON_REQUIRED": 422,
          "CLARIFICATION_ROUND_INVALID": 409,
          "CLARIFICATION_ROUND_LIMIT_REACHED": 409,
          "DEEP_CONFIRMATION_REQUIRED": 409,
          "DEPENDENCY_UNAVAILABLE": 503,
          "FORBIDDEN": 403,
          "IDEMPOTENCY_CONFLICT": 409,
          "PERMISSION_CHANGED": 403,
          "QUEUE_UNAVAILABLE": 503,
          "RESOURCE_NOT_FOUND": 404,
          "RESULT_NOT_READY": 409,
          "RESULT_QUALITY_BLOCKED": 409,
          "RISK_ACCEPTANCE_INVALID": 422,
          "STALE_TARGET": 409,
          "TRACEABILITY_INCOMPLETE": 409,
          "VALIDATION_ERROR": 422,
          "VERSION_CONFLICT": 409
        },
        "x-implementation-status": "implemented-p1-runtime-r4-candidate",
        "x-permission": {
          "admin_bypass": false,
          "allowed_project_roles": [
            "owner",
            "reviewer",
            "implementer",
            "tester"
          ]
        },
        "x-permissions": {
          "admin_bypass": false,
          "allowed_project_roles": [
            "owner",
            "reviewer",
            "implementer",
            "tester"
          ]
        }
      },
      "post": {
        "description": "Persists raw_input unchanged at requirement_version.content_json.raw_input. Subsequent Requirement Versions inherit raw_input and cannot modify it.",
        "operationId": "createRequirement",
        "parameters": [
          {
            "in": "path",
            "name": "version_id",
            "required": true,
            "schema": {
              "type": "string"
            }
          },
          {
            "$ref": "#/components/parameters/IdempotencyKey"
          }
        ],
        "requestBody": {
          "content": {
            "application/json": {
              "schema": {
                "$ref": "#/components/schemas/CreateRequirementRequest"
              }
            }
          },
          "required": true
        },
        "responses": {
          "200": {
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/RequirementResponse"
                }
              }
            },
            "description": "Successful response"
          },
          "401": {
            "$ref": "#/components/responses/Sprint2Error401"
          },
          "403": {
            "$ref": "#/components/responses/Sprint2Error403"
          },
          "404": {
            "$ref": "#/components/responses/Sprint2Error404"
          },
          "409": {
            "$ref": "#/components/responses/Sprint2Error409"
          },
          "422": {
            "$ref": "#/components/responses/Sprint2Error422"
          },
          "429": {
            "$ref": "#/components/responses/Sprint2Error429"
          },
          "503": {
            "$ref": "#/components/responses/Sprint2Error503"
          }
        },
        "security": [
          {
            "bearerAuth": []
          }
        ],
        "tags": [
          "requirements"
        ],
        "x-command-semantics": {
          "actor": "owner",
          "admin_bypass": false,
          "queue_unavailable_no_false_queued": false
        },
        "x-contract-phase": "sprint-2-p1-runtime-r4-candidate",
        "x-error-http-mapping": {
          "AI_QUOTA_EXCEEDED": 429,
          "AI_TASK_NOT_CANCELLABLE": 409,
          "AI_TASK_NOT_RETRYABLE": 409,
          "AUTH_REQUIRED": 401,
          "CLARIFICATION_MODE_REASON_REQUIRED": 422,
          "CLARIFICATION_ROUND_INVALID": 409,
          "CLARIFICATION_ROUND_LIMIT_REACHED": 409,
          "DEEP_CONFIRMATION_REQUIRED": 409,
          "DEPENDENCY_UNAVAILABLE": 503,
          "FORBIDDEN": 403,
          "IDEMPOTENCY_CONFLICT": 409,
          "PERMISSION_CHANGED": 403,
          "QUEUE_UNAVAILABLE": 503,
          "RESOURCE_NOT_FOUND": 404,
          "RESULT_NOT_READY": 409,
          "RESULT_QUALITY_BLOCKED": 409,
          "RISK_ACCEPTANCE_INVALID": 422,
          "STALE_TARGET": 409,
          "TRACEABILITY_INCOMPLETE": 409,
          "VALIDATION_ERROR": 422,
          "VERSION_CONFLICT": 409
        },
        "x-idempotency-scope": "actor+endpoint+request-hash",
        "x-implementation-status": "implemented-p1-runtime-r4-candidate",
        "x-permission": {
          "admin_bypass": false,
          "allowed_project_roles": [
            "owner"
          ]
        },
        "x-permissions": {
          "admin_bypass": false,
          "allowed_project_roles": [
            "owner"
          ]
        },
        "x-r4-decisions": [
          "permission",
          "transaction",
          "event"
        ],
        "x-raw-input-policy": {
          "immutable_across_versions": true,
          "persisted_path": "requirement_version.content_json.raw_input",
          "ref_content_hash": "sha256(utf8(raw_input))"
        },
        "x-transaction": {
          "atomic": true,
          "audit_required": true,
          "boundary": "identity+initial_version+audit+idempotency+business_outbox",
          "outbox_required": true,
          "task_creation_decoupled": false
        },
        "x-transaction-boundary": "identity+initial_version+audit+idempotency+business_outbox"
      }
    },
...
    "/api/v1/requirement-versions/{version_id}": {
      "patch": {
        "description": "Creates a new immutable Requirement Version and never overwrites the addressed version; raw_input is inherited and cannot be modified.",
        "operationId": "reviseRequirementVersion",
        "parameters": [
          {
            "in": "path",
            "name": "version_id",
            "required": true,
            "schema": {
              "type": "string"
            }
          }
        ],
        "requestBody": {
          "content": {
            "application/json": {
              "schema": {
                "$ref": "#/components/schemas/ReviseRequirementVersionRequest"
              }
            }
          },
          "required": true
        },
        "responses": {
          "200": {
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/RequirementVersionResponse"
                }
              }
            },
            "description": "Successful response"
          },
          "401": {
            "$ref": "#/components/responses/Sprint2Error401"
          },
          "403": {
            "$ref": "#/components/responses/Sprint2Error403"
          },
          "404": {
            "$ref": "#/components/responses/Sprint2Error404"
          },
          "409": {
            "$ref": "#/components/responses/Sprint2Error409"
          },
          "422": {
            "$ref": "#/components/responses/Sprint2Error422"
          },
          "429": {
            "$ref": "#/components/responses/Sprint2Error429"
          },
          "503": {
            "$ref": "#/components/responses/Sprint2Error503"
          }
        },
        "security": [
          {
            "bearerAuth": []
          }
        ],
        "tags": [
          "requirements"
        ],
        "x-command-semantics": {
          "actor": "owner",
          "admin_bypass": false,
          "queue_unavailable_no_false_queued": false
        },
        "x-contract-phase": "sprint-2-p1-runtime-r4-candidate",
        "x-error-http-mapping": {
          "AI_QUOTA_EXCEEDED": 429,
          "AI_TASK_NOT_CANCELLABLE": 409,
          "AI_TASK_NOT_RETRYABLE": 409,
          "AUTH_REQUIRED": 401,
          "CLARIFICATION_MODE_REASON_REQUIRED": 422,
          "CLARIFICATION_ROUND_INVALID": 409,
          "CLARIFICATION_ROUND_LIMIT_REACHED": 409,
          "DEEP_CONFIRMATION_REQUIRED": 409,
          "DEPENDENCY_UNAVAILABLE": 503,
          "FORBIDDEN": 403,
          "IDEMPOTENCY_CONFLICT": 409,
          "PERMISSION_CHANGED": 403,
          "QUEUE_UNAVAILABLE": 503,
          "RESOURCE_NOT_FOUND": 404,
          "RESULT_NOT_READY": 409,
          "RESULT_QUALITY_BLOCKED": 409,
          "RISK_ACCEPTANCE_INVALID": 422,
          "STALE_TARGET": 409,
          "TRACEABILITY_INCOMPLETE": 409,
          "VALIDATION_ERROR": 422,
          "VERSION_CONFLICT": 409
        },
        "x-implementation-status": "implemented-p1-runtime-r4-candidate",
        "x-permission": {
          "admin_bypass": false,
          "allowed_project_roles": [
            "owner"
          ]
        },
        "x-permissions": {
          "admin_bypass": false,
          "allowed_project_roles": [
            "owner"
          ]
        },
        "x-r4-decisions": [
          "permission",
          "idempotency",
          "transaction",
          "event"
        ],
        "x-transaction": {
          "atomic": true,
          "audit_required": true,
          "boundary": "new_version+audit+business_outbox",
          "outbox_required": true,
          "task_creation_decoupled": false
        },
        "x-transaction-boundary": "new_version+audit+business_outbox"
      }
    },
    "/api/v1/requirement-versions/{version_id}/clarification-answers": {
      "post": {
        "description": "Every answer submission creates a new immutable Requirement Version.",
        "operationId": "submitRequirementClarificationAnswers",
        "parameters": [
          {
            "in": "path",
            "name": "version_id",
            "required": true,
            "schema": {
              "type": "string"
            }
          },
          {
            "$ref": "#/components/parameters/IdempotencyKey"
          }
        ],
        "requestBody": {
          "content": {
            "application/json": {
              "schema": {
                "$ref": "#/components/schemas/SubmitClarificationAnswersRequest"
              }
            }
          },
          "required": true
        },
        "responses": {
          "200": {
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/ClarificationAnswerResponse"
                }
              }
            },
            "description": "Successful response"
          },
          "401": {
            "$ref": "#/components/responses/Sprint2Error401"
          },
          "403": {
            "$ref": "#/components/responses/Sprint2Error403"
          },
          "404": {
            "$ref": "#/components/responses/Sprint2Error404"
          },
          "409": {
            "$ref": "#/components/responses/Sprint2Error409"
          },
          "422": {
            "$ref": "#/components/responses/Sprint2Error422"
          },
          "429": {
            "$ref": "#/components/responses/Sprint2Error429"
          },
          "503": {
            "$ref": "#/components/responses/Sprint2Error503"
          }
        },
        "security": [
          {
            "bearerAuth": []
          }
        ],
        "tags": [
          "requirements"
        ],
        "x-command-semantics": {
          "actor": "owner",
          "admin_bypass": false,
          "queue_unavailable_no_false_queued": false
        },
        "x-contract-phase": "sprint-2-p1-runtime-r4-candidate",
        "x-error-http-mapping": {
          "AI_QUOTA_EXCEEDED": 429,
          "AI_TASK_NOT_CANCELLABLE": 409,
          "AI_TASK_NOT_RETRYABLE": 409,
          "AUTH_REQUIRED": 401,
          "CLARIFICATION_MODE_REASON_REQUIRED": 422,
          "CLARIFICATION_ROUND_INVALID": 409,
          "CLARIFICATION_ROUND_LIMIT_REACHED": 409,
          "DEEP_CONFIRMATION_REQUIRED": 409,
          "DEPENDENCY_UNAVAILABLE": 503,
          "FORBIDDEN": 403,
          "IDEMPOTENCY_CONFLICT": 409,
          "PERMISSION_CHANGED": 403,
          "QUEUE_UNAVAILABLE": 503,
          "RESOURCE_NOT_FOUND": 404,
          "RESULT_NOT_READY": 409,
          "RESULT_QUALITY_BLOCKED": 409,
          "RISK_ACCEPTANCE_INVALID": 422,
          "STALE_TARGET": 409,
          "TRACEABILITY_INCOMPLETE": 409,
          "VALIDATION_ERROR": 422,
          "VERSION_CONFLICT": 409
        },
        "x-idempotency-scope": "actor+endpoint+request-hash",
        "x-implementation-status": "implemented-p1-runtime-r4-candidate",
        "x-permission": {
          "admin_bypass": false,
          "allowed_project_roles": [
            "owner"
          ]
        },
        "x-permissions": {
          "admin_bypass": false,
          "allowed_project_roles": [
            "owner"
          ]
        },
        "x-r4-decisions": [
          "permission",
          "error_codes",
          "transaction",
          "event",
          "ai_task_acceptance"
        ],
        "x-transaction": {
          "atomic": true,
          "audit_required": true,
          "boundary": "answer_version+audit+business_outbox; ai_task_creation_decoupled",
          "outbox_required": true,
          "task_creation_decoupled": true
        },
        "x-transaction-boundary": "answer_version+audit+business_outbox; ai_task_creation_decoupled"
      }
    },
    "/api/v1/requirement-versions/{version_id}:confirm": {
      "post": {
        "operationId": "confirmRequirementVersion",
        "parameters": [
          {
            "in": "path",
            "name": "version_id",
            "required": true,
            "schema": {
              "type": "string"
            }
          },
          {
            "$ref": "#/components/parameters/IdempotencyKey"
          }
        ],
        "requestBody": {
          "content": {
            "application/json": {
              "schema": {
                "$ref": "#/components/schemas/ConfirmRequirementVersionRequest"
              }
            }
          },
          "required": true
        },
        "responses": {
          "200": {
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/ConfirmRequirementResponse"
                }
              }
            },
            "description": "Successful response"
          },
          "401": {
            "$ref": "#/components/responses/Sprint2Error401"
          },
          "403": {
            "$ref": "#/components/responses/Sprint2Error403"
          },
          "404": {
            "$ref": "#/components/responses/Sprint2Error404"
          },
          "409": {
            "$ref": "#/components/responses/Sprint2Error409"
          },
          "422": {
            "$ref": "#/components/responses/Sprint2Error422"
          },
          "429": {
            "$ref": "#/components/responses/Sprint2Error429"
          },
          "503": {
            "$ref": "#/components/responses/Sprint2Error503"
          }
        },
        "security": [
          {
            "bearerAuth": []
          }
        ],
        "tags": [
          "requirements"
        ],
        "x-command-semantics": {
          "actor": "owner",
          "admin_bypass": false,
          "queue_unavailable_no_false_queued": false
        },
        "x-contract-phase": "sprint-2-p1-runtime-r4-candidate",
        "x-error-http-mapping": {
          "AI_QUOTA_EXCEEDED": 429,
          "AI_TASK_NOT_CANCELLABLE": 409,
          "AI_TASK_NOT_RETRYABLE": 409,
          "AUTH_REQUIRED": 401,
          "CLARIFICATION_MODE_REASON_REQUIRED": 422,
          "CLARIFICATION_ROUND_INVALID": 409,
          "CLARIFICATION_ROUND_LIMIT_REACHED": 409,
          "DEEP_CONFIRMATION_REQUIRED": 409,
          "DEPENDENCY_UNAVAILABLE": 503,
          "FORBIDDEN": 403,
          "IDEMPOTENCY_CONFLICT": 409,
          "PERMISSION_CHANGED": 403,
          "QUEUE_UNAVAILABLE": 503,
          "RESOURCE_NOT_FOUND": 404,
          "RESULT_NOT_READY": 409,
          "RESULT_QUALITY_BLOCKED": 409,
          "RISK_ACCEPTANCE_INVALID": 422,
          "STALE_TARGET": 409,
          "TRACEABILITY_INCOMPLETE": 409,
          "VALIDATION_ERROR": 422,
          "VERSION_CONFLICT": 409
        },
        "x-idempotency-scope": "actor+endpoint+request-hash",
        "x-implementation-status": "implemented-p1-runtime-r4-candidate",
        "x-permission": {
          "admin_bypass": false,
          "allowed_project_roles": [
            "owner"
          ]
        },
        "x-permissions": {
          "admin_bypass": false,
          "allowed_project_roles": [
            "owner"
          ]
        },
        "x-r4-decisions": [
          "permission",
          "gate_result_enum",
          "transaction",
          "event"
        ],
        "x-transaction": {
          "atomic": true,
          "audit_required": true,
          "boundary": "effective_version+pointer+audit+business_outbox",
          "outbox_required": true,
          "task_creation_decoupled": false
        },
        "x-transaction-boundary": "effective_version+pointer+audit+business_outbox"
      }
    },
    "/api/v1/requirement-versions/{version_id}:set-clarification-mode": {
      "post": {
        "operationId": "setRequirementClarificationMode",
        "parameters": [
          {
            "in": "path",
            "name": "version_id",
            "required": true,
            "schema": {
              "type": "string"
            }
          },
          {
            "$ref": "#/components/parameters/IdempotencyKey"
          }
        ],
        "requestBody": {
          "content": {
            "application/json": {
              "schema": {
                "$ref": "#/components/schemas/SetClarificationModeRequest"
              }
            }
          },
          "required": true
        },
        "responses": {
          "200": {
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/RequirementVersionResponse"
                }
              }
            },
            "description": "Successful response"
          },
          "401": {
            "$ref": "#/components/responses/Sprint2Error401"
          },
          "403": {
            "$ref": "#/components/responses/Sprint2Error403"
          },
          "404": {
            "$ref": "#/components/responses/Sprint2Error404"
          },
          "409": {
            "$ref": "#/components/responses/Sprint2Error409"
          },
          "422": {
            "$ref": "#/components/responses/Sprint2Error422"
          },
          "429": {
            "$ref": "#/components/responses/Sprint2Error429"
          },
          "503": {
            "$ref": "#/components/responses/Sprint2Error503"
          }
        },
        "security": [
          {
            "bearerAuth": []
          }
        ],
        "tags": [
          "requirements"
        ],
        "x-command-semantics": {
          "actor": "owner",
          "admin_bypass": false,
          "queue_unavailable_no_false_queued": false
        },
        "x-contract-phase": "sprint-2-p1-runtime-r4-candidate",
        "x-error-http-mapping": {
          "AI_QUOTA_EXCEEDED": 429,
          "AI_TASK_NOT_CANCELLABLE": 409,
          "AI_TASK_NOT_RETRYABLE": 409,
          "AUTH_REQUIRED": 401,
          "CLARIFICATION_MODE_REASON_REQUIRED": 422,
          "CLARIFICATION_ROUND_INVALID": 409,
          "CLARIFICATION_ROUND_LIMIT_REACHED": 409,
          "DEEP_CONFIRMATION_REQUIRED": 409,
          "DEPENDENCY_UNAVAILABLE": 503,
          "FORBIDDEN": 403,
          "IDEMPOTENCY_CONFLICT": 409,
          "PERMISSION_CHANGED": 403,
          "QUEUE_UNAVAILABLE": 503,
          "RESOURCE_NOT_FOUND": 404,
          "RESULT_NOT_READY": 409,
          "RESULT_QUALITY_BLOCKED": 409,
          "RISK_ACCEPTANCE_INVALID": 422,
          "STALE_TARGET": 409,
          "TRACEABILITY_INCOMPLETE": 409,
          "VALIDATION_ERROR": 422,
          "VERSION_CONFLICT": 409
        },
        "x-idempotency-scope": "actor+endpoint+request-hash",
        "x-implementation-status": "implemented-p1-runtime-r4-candidate",
        "x-permission": {
          "admin_bypass": false,
          "allowed_project_roles": [
            "owner"
          ]
        },
        "x-permissions": {
          "admin_bypass": false,
          "allowed_project_roles": [
            "owner"
          ]
        },
        "x-r4-decisions": [
          "permission",
          "new_version_semantics",
          "transaction",
          "event"
        ],
        "x-transaction": {
          "atomic": true,
          "audit_required": true,
          "boundary": "new_version+audit+business_outbox",
          "outbox_required": true,
          "task_creation_decoupled": false
        },
        "x-transaction-boundary": "new_version+audit+business_outbox"
      }
    },
    "/api/v1/requirements/{requirement_id}": {
      "get": {
        "operationId": "getRequirement",
        "parameters": [
          {
            "in": "path",
            "name": "requirement_id",
            "required": true,
            "schema": {
              "type": "string"
            }
          }
        ],
        "responses": {
          "200": {
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/RequirementResponse"
                }
              }
            },
            "description": "Successful response"
          },
          "401": {
            "$ref": "#/components/responses/Sprint2Error401"
          },
          "403": {
            "$ref": "#/components/responses/Sprint2Error403"
          },
          "404": {
            "$ref": "#/components/responses/Sprint2Error404"
          },
          "409": {
            "$ref": "#/components/responses/Sprint2Error409"
          },
          "422": {
            "$ref": "#/components/responses/Sprint2Error422"
          },
          "429": {
            "$ref": "#/components/responses/Sprint2Error429"
          },
          "503": {
            "$ref": "#/components/responses/Sprint2Error503"
          }
        },
        "security": [
          {
            "bearerAuth": []
          }
        ],
        "tags": [
          "requirements"
        ],
        "x-command-semantics": {
          "actor": "project_member_with_target_read_permission",
          "admin_bypass": false,
          "queue_unavailable_no_false_queued": false
        },
        "x-contract-phase": "sprint-2-p1-runtime-r4-candidate",
        "x-error-http-mapping": {
          "AI_QUOTA_EXCEEDED": 429,
          "AI_TASK_NOT_CANCELLABLE": 409,
          "AI_TASK_NOT_RETRYABLE": 409,
          "AUTH_REQUIRED": 401,
          "CLARIFICATION_MODE_REASON_REQUIRED": 422,
          "CLARIFICATION_ROUND_INVALID": 409,
          "CLARIFICATION_ROUND_LIMIT_REACHED": 409,
          "DEEP_CONFIRMATION_REQUIRED": 409,
          "DEPENDENCY_UNAVAILABLE": 503,
          "FORBIDDEN": 403,
          "IDEMPOTENCY_CONFLICT": 409,
          "PERMISSION_CHANGED": 403,
          "QUEUE_UNAVAILABLE": 503,
          "RESOURCE_NOT_FOUND": 404,
          "RESULT_NOT_READY": 409,
          "RESULT_QUALITY_BLOCKED": 409,
          "RISK_ACCEPTANCE_INVALID": 422,
          "STALE_TARGET": 409,
          "TRACEABILITY_INCOMPLETE": 409,
          "VALIDATION_ERROR": 422,
          "VERSION_CONFLICT": 409
        },
        "x-implementation-status": "implemented-p1-runtime-r4-candidate",
        "x-permission": {
          "admin_bypass": false,
          "allowed_project_roles": [
            "owner",
            "reviewer",
            "implementer",
            "tester"
          ]
        },
        "x-permissions": {
          "admin_bypass": false,
          "allowed_project_roles": [
            "owner",
            "reviewer",
            "implementer",
            "tester"
          ]
        }
      }
    },
...
  },
...
}